+++ src/test/java/com/github/jaypatel/hangman/HangmanTest.java
@@ -1,0 +1,15 @@
+package com.github.jaypatel.hangman;
+
+import static org.junit.Assert.assertArrayEquals;
+
+import org.junit.Test;
+
+public class HangmanTest {
+
+    @Test
+    public void testing() {
+        //Setup
+        
+    }
+
+}