+++ Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 0.1.1
+### Added 
+- Basic java docs
+- Added the ability to do Junit testing
+
+### Changes
+- Some minor code refactoring which includes the startGame() method
+
+
 ## Version 0.1.0
 ### Added
 - Mavenized the project